requirements.txt
scikit-learn == 1.2.2
Pillow == 9.5.0
streamlit == 1.21.0
google-api-python-client == 2.86.0
 pandas == 1.5.3
 scipy == 1.10.1
 numpy >= 1.17.3
 gunicorn == 20.1.0
 Flask == 2.2.3
 itsdangerous == 2.1.2
Jinja2 == 3.1.2
 MarkupSafe == 2.1.2
Werkzeug == 2.2.3
 matplotlib == 3.7.1
 seaborn == 0.12.2
tensorflow == 2.12.0
keras	==	2.12.0
 protobuf==3.20.3
